Sorter tester
